Update stock tickers and refine data analysis scripts
The commit redefines the stock tickers used for analysis in the matplotlib, scipy, and statsmodels examples. It also adds more specificity in the correlation calculation in the matplotlib script. The import of pandas from the scipy example is removed. Furthermore, the S&P 500 ticker list gets updated and the yfinance example script downloads data for DJI in addition to SPYI. Lastly, the requirements.txt file includes scipy.